Split out some functions, and make script behave more unix-y
If the -a option is used, then the script will use the pwd as the source
directory if you don't give it one. You no longer need to pass in 'pwd'
as the last argument to accomplish this. This makes the script more
composable with pipes and other programs and such, as it no longer tries
to evaluate that last argument.